appointment module crud apis added and routes re-structured
