+++ nodejs_fhir_poc/.gitignore
@@ -1,2 +1,2 @@
-package-lock.json
 node_modules
+package-lock.json
