--- nodejs_fhir_poc/FHIR-Poc.postman_collection.json
@@ -1,109 +1,0 @@
-{
-	"info": {
-		"_postman_id": "59f5bde5-832a-49e2-8110-b195e757936a",
-		"name": "FHIR-Poc",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "2753995"
-	},
-	"item": [
-		{
-			"name": "Create",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/fhir",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"fhir"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "GET",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:8080/fhir/Bundle",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "8080",
-					"path": [
-						"fhir",
-						"Bundle"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "GETByID",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/fhir/3",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"fhir",
-						"3"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Update",
-			"request": {
-				"method": "PATCH",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/fhir/3",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"fhir",
-						"3"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Delete",
-			"request": {
-				"method": "DELETE",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/fhir/2",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"fhir",
-						"2"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
