+++ nodejs_fhir_poc/app.js
@@ -6,6 +6,8 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
+var appointmentRouter = require('./routes/appointment');
+var adverseeventRouter = require('./routes/AdverseEvent');
 
 var app = express();
 
@@ -21,6 +23,8 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/fhir/adverseevent', adverseeventRouter);
+app.use('/fhir/appointment', appointmentRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
